UI: Create Interface and Base Class for Drilldown
