+++ components/ILIAS/UI/src/templates/default/Menu/tpl.drilldown.html
@@ -1,6 +1,11 @@
-<div class="il-drilldown" id="{ID}">
+<div class="il-drilldown{CATEGORY}" id="{ID}">
     <header class="show-title show-backnav">
+        <!-- BEGIN title -->
         <h2>{TITLE}</h2>
+        <!-- END title -->
+        <!-- BEGIN filter -->
+        {FILTER}
+        <!-- END title -->
         <div class="backnav">
             {BACKNAV}
         </div>
